@@ -5,18 +5,18 @@
   "authors": [
     {
       "name": "William Eldridge",
-      "homepage": "http://wigedev.net"
+      "homepage": "https://wigedev.net"
     }
   ],
   "require": {
     "ext-pdo": "*",
     "jasperfw/core": "^1.0",
     "jasperfw/validator": "^1.0",
-    "php": "^7.2",
+    "php": "^8.1",
     "psr/log": "1.0.*"
   },
   "require-dev": {
-    "phpunit/phpunit": "~8.0"
+    "phpunit/phpunit": "~9.0"
   },
   "autoload": {
     "psr-4": {
